+++ HISTORY.rst
@@ -4,6 +4,7 @@ Version 0.2.2 - XXXX-XX-XX
 ==========================
 
 - #18: isurl() validator.
+- #19: isip4(), isip6() and isip46() validators.
 
 Version 0.2.1 - 2015-07-28
 ==========================